退会ページを質問ページと分けた
