+++ src/components/organisms/Footor.tsx
@@ -1,5 +1,19 @@
+import Link from "next/link";
+
 function Footer() {
-  return <footer className="h-14 bg-stone-200">利用規約</footer>;
+  return (
+    <div className="space-x-10 text-center">
+      <Link href="/">
+        <a>利用規約</a>
+      </Link>
+      <Link href="/">
+        <a>プライバシーポリシー</a>
+      </Link>
+      <Link href="/delete-account">
+        <a>退会方法</a>
+      </Link>
+    </div>
+  );
 }
 
 export default Footer;
