+++ src/components/organisms/Questions.tsx
@@ -1,69 +1,17 @@
-import { auth, provider } from "auth/firebase";
-import { deleteUser, reauthenticateWithPopup } from "firebase/auth";
-import Button from "components/atoms/Button";
 import Heading from "components/atoms/Heading";
-import router from "next/router";
-import { useGetUsersUserId } from "api/users/users";
-import { useDeleteAuthentication } from "api/default/default";
 
 function Questions() {
-  const { data } = useGetUsersUserId();
-  const user = auth.currentUser;
-
-  const reauthenticate = () => {
-    user &&
-      reauthenticateWithPopup(user, provider)
-        .then(() => {})
-        .catch((error) => {
-          console.log(error);
-        });
-  };
-
-  const deleteServerSideData = useDeleteAuthentication();
-
-  const deleteAccount = () =>
-    user &&
-    deleteUser(user)
-      .then(() => {
-        data && deleteServerSideData.mutate({ userId: data.id });
-        console.log("An account has been deleted.");
-        alert(
-          "退会処理が完了しました。ご利用いただきましてありがとうございました。"
-        );
-      })
-      .catch((error) => {
-        if (error.code === "auth/requires-recent-login") {
-          reauthenticate();
-          alert(
-            "アカウント認証に失敗したため処理が実行できませんでした。再度お試しください。"
-          );
-        } else {
-          console.log(error);
-        }
-      });
-
-  const handleClick = () => {
-    const confirmation = confirm(
-      "全てのデータが消えてしまいますが退会しますか？"
-    );
-    confirmation ? deleteAccount() : console.log("Canceled.");
-  };
   return (
     <div className="flex justify-center mt-6">
       <div>
         <Heading text="よくある質問" />
-        <h3 className="my-3">退会方法を教えてください。</h3>
-        <p>退会処理は以下のボタンから行うことができます。</p>
+        <h3>為替レートは更新されているのでしょうか？</h3>
         <p>
-          ※
-          退会処理を行うとこれまでのデータが全て消えてしまいますのでご注意ください。
+          為替レートは1日1回最新情報を取得しております。本サービス開始以前の日付や未来の日付など、取得できない日のレートに対してはその日の最新レートを代わりに表示しております。
         </p>
         <p>
-          ※ 退会処理を行う際にGooleアカウントの再認証が必要な場合がございます。
+          月末に取得したレートをその月の最終レートとして確定し、その月の給料を計算します。
         </p>
-        <div className="my-6 text-center">
-          <Button text="退会する" onClick={handleClick} />
-        </div>
       </div>
     </div>
   );
