+++ src/pages/delete-account.tsx
@@ -1,0 +1,10 @@
+import DeleteAccount from "components/organisms/DeleteAccount";
+import Main from "components/templates/Main";
+
+export default function Account() {
+  return (
+    <Main>
+      <DeleteAccount />
+    </Main>
+  );
+}
